refactor: 레이싱 진행 관련기능 분리
